Move pr_error to common utils and fix implicit declaration
- Moved `pr_error` from `installer.c` to `common/utils.c` and `common/utils.h`.
- Added necessary includes for `pr_error` in `common/utils.c`.
- This fixes the compilation error in `gui.c` where `pr_error` was implicitly declared.
